requirements.txt
fastapi==0.115.12
httptools==0.6.4
paho-mqtt==2.1.0
psutil==7.2.2
python-dotenv==1.2.2
uvicorn==0.34.2
uvloop==0.21.0
